+++ src/pluginManager.ts
@@ -4,7 +4,7 @@
  * @file plugins.ts
  * @author Luca Liguori
  * @created 2024-07-14
- * @version 1.3.3
+ * @version 1.3.4
  * @license Apache-2.0
  *
  * Copyright 2024, 2025, 2026 Luca Liguori.
@@ -75,11 +75,12 @@ export class PluginManager extends EventEmitter<PluginManagerEvents> {
   }
 
   destroy(): void {
+    this.server.off('broadcast_message', this.msgHandler.bind(this));
     this.server.close();
   }
 
   private async msgHandler(msg: WorkerMessage): Promise<void> {
-    if (this.server.isWorkerRequest(msg) && (msg.dst === 'all' || msg.dst === 'plugins')) {
+    if (this.server.isWorkerRequest(msg)) {
       if (this.verbose) this.log.debug(`Received request message ${CYAN}${msg.type}${db} from ${CYAN}${msg.src}${db}: ${debugStringify(msg)}${db}`);
       switch (msg.type) {
         case 'get_log_level':
@@ -491,6 +492,7 @@ export class PluginManager extends EventEmitter<PluginManagerEvents> {
 
   /**
    * Resolves the name of a plugin by loading and parsing its package.json file.
+   * It will first try to resolve the path as is, then in the global modules directory, and finally in the matterbridge plugin directory.
    *
    * @param {string} nameOrPath - The name of the plugin or the path to the plugin's package.json file.
    * @returns {Promise<string | null>} A promise that resolves to the path of the plugin's package.json file or null if it could not be resolved.
@@ -504,7 +506,7 @@ export class PluginManager extends EventEmitter<PluginManagerEvents> {
     let packageJsonPath = path.resolve(nameOrPath);
     this.log.debug(`Resolving plugin path ${plg}${packageJsonPath}${db}`);
 
-    // Check if the package.json file exists
+    // Check if the package.json file exists at the specified path or next try in the global modules directory
     try {
       await promises.access(packageJsonPath);
     } catch {
@@ -512,8 +514,16 @@ export class PluginManager extends EventEmitter<PluginManagerEvents> {
       packageJsonPath = path.join(this.matterbridge.globalModulesDirectory, nameOrPath);
       this.log.debug(`Trying at ${plg}${packageJsonPath}${db}`);
     }
+    // Check if the package.json file exists at the global modules directory or next try in the matterbridge plugin directory
     try {
-      // Load the package.json of the plugin
+      await promises.access(packageJsonPath);
+    } catch {
+      this.log.debug(`Package.json not found at ${plg}${packageJsonPath}${db}`);
+      packageJsonPath = path.join(this.matterbridge.matterbridgePluginDirectory, nameOrPath);
+      this.log.debug(`Trying at ${plg}${packageJsonPath}${db}`);
+    }
+    try {
+      // Load the package.json of the plugin or fails if even not found in the matterbridge plugin directory
       const packageJson = JSON.parse(await promises.readFile(packageJsonPath, 'utf8'));
 
       // Check for main issues
@@ -1263,7 +1273,6 @@ export class PluginManager extends EventEmitter<PluginManagerEvents> {
   async loadConfig(plugin: Plugin): Promise<PlatformConfig> {
     const { default: path } = await import('node:path');
     const { promises } = await import('node:fs');
-    const { shelly_config, somfytahoma_config, zigbee2mqtt_config } = await import('./defaultConfigSchema.js');
     const configFile = path.join(this.matterbridge.matterbridgeDirectory, `${plugin.name}.config.json`);
     const defaultConfigFile = plugin.path.replace('package.json', `${plugin.name}.config.json`);
 
@@ -1291,12 +1300,7 @@ export class PluginManager extends EventEmitter<PluginManagerEvents> {
           this.log.debug(`Loaded default config file ${defaultConfigFile} for plugin ${plg}${plugin.name}${db}.`);
         } catch (_err) {
           this.log.debug(`Default config file ${defaultConfigFile} for plugin ${plg}${plugin.name}${db} does not exist, creating new config file...`);
-          // TODO: Remove this when all these plugins have their own default config file
-          // istanbul ignore next if
-          if (plugin.name === 'matterbridge-zigbee2mqtt') config = zigbee2mqtt_config;
-          else if (plugin.name === 'matterbridge-somfy-tahoma') config = somfytahoma_config;
-          else if (plugin.name === 'matterbridge-shelly') config = shelly_config;
-          else config = { name: plugin.name, type: plugin.type, version: '1.0.0', debug: false, unregisterOnShutdown: false };
+          config = { name: plugin.name, type: plugin.type, version: '1.0.0', debug: false, unregisterOnShutdown: false };
         }
         try {
           await promises.writeFile(configFile, JSON.stringify(config, null, 2), 'utf8');